requirements.txt

fastapi>=0.100.0,<0.105.0
uvicorn[standard]>=0.20.0,<0.25.0
python-multipart>=0.0.15,<0.1.0

langchain-groq
httpx>=0.24.0,<0.26.0
aiohttp>=3.8.0,<4.0.0

python-dotenv>=1.0.0,<2.0.0
requests>=2.30.0,<3.0.0
pydantic>=2.0.0,<3.0.0
pdfplumber>=0.9.0,<0.11.0
PyMuPDF>=1.24.0,<1.26.0

pinecone-client>=3.0.0,<4.0.0

langchain-core==0.3.69
langchain==0.3.26
langchain-community==0.3.27
langchain-huggingface
langchain-text-splitters>=0.2.0,<0.4.0
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.0+cpu # Replace with your required version
sentence-transformers
typing-extensions>=4.8.0,<5.0.0
packaging>=23.0,<25.0
tenacity>=8.0.0,<10.0.0
PyYAML>=6.0.0,<7.0.0

gunicorn>=21.0.0,<22.0.0
